+++ R/beta-binomial-distribution.R
@@ -68,6 +68,7 @@
 #' hist(x, 100, freq = FALSE)
 #' lines(xx-0.5, dbbinom(xx, 1000, 5, 13), col = "red")
 #' hist(pbbinom(x, 1000, 5, 13))
+#' xx <- seq(0, 1000, by = 0.1)
 #' plot(ecdf(x))
 #' lines(xx, pbbinom(xx, 1000, 5, 13), col = "red", lwd = 2)
 #'