+++ R/laplace-distribution.R
@@ -36,11 +36,11 @@
 #' Quantile function
 #' \deqn{ F^{-1}(p) = \left\{\begin{array}{ll}
 #' \mu + \sigma \log(2p)     & p < 0.5 \\
-#' \mu + \sigma \log(2(1-p)) & p \geq 0.5
+#' \mu - \sigma \log(2(1-p)) & p \geq 0.5
 #' \end{array}\right.
 #' }{
 #' F^-1(p) = [if p < 0.5:] \mu + \sigma * log(2*p)
-#'           [else:] \mu + \sigma * log(2*(1-p))
+#'           [else:] \mu - \sigma * log(2*(1-p))
 #' }
 #'
 #' @references